chore: rename interface to types
